requirements.txt
streamlit
numpy
Pillow
numpy-stl
plotly
